LayoutTests/platform/chromium/accessibility/other-aria-attribute-change-sends-notification-expected.txt

This test ensures that a change to any ARIA attribute, not just a state, sends a notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Got AriaAttributeChanged notification on aria-busy
Got AriaAttributeChanged notification on aria-disabled
Got AriaAttributeChanged notification on aria-readonly
Got AriaAttributeChanged notification on aria-required
All notifications received successfully.
PASS successfullyParsed is true

TEST COMPLETE
